Tempis fugit!      YOU NOW REQUIRE JAVA 8 TO SUCCESSFULLY COMPILE THIS REPO.
Apologies to the very few unable to get Java 8, but for most of the world, Java 8 is in
late beta, due for birth in about 6 weeks. The world moves on, and so must this repo.
I am polishing the release of the Java Cookbook Third Edition, and this has to be ready.