Start, and finish, FFI function example...
We pass a string into a (trivial) datastore written in C, and retrieve the
value with that index, to prove that we can pass strings into C code and
retrieve (the same) string back from the C code.
Add the Java code, the C code, and a makefile since the C code has to
be installed as a shared library.